Make the Scene model the single source of truth for Fixtures
This commit removes the duplicated fixture lists and caches and makes the
SceneController defer to the Scene model for storing Fixture objects. To
simplify the refactor, it adds a SceneController reference to the Scene
model. The real solution is to store controller/UI-specific data in a separate
FixtureUI model.

This commit also breaks adding new strands to scenes, but that will be fixed in
the next commit.